viz-demo/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

#
pandas
plotly
kaleido==0.1.0post1  # Render images (note: later version have issues on windows 10+)

#
python-pptx # power point library